+++ src/zapv2/wappalyzer.py
@@ -2,7 +2,7 @@
 #
 # ZAP is an HTTP/HTTPS proxy for assessing web application security.
 #
-# Copyright 2016 the ZAP development team
+# Copyright 2020 the ZAP development team
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,27 +22,30 @@ This file was automatically generated.
 import six
 
 
-class openapi(object):
+class wappalyzer(object):
 
     def __init__(self, zap):
         self.zap = zap
 
-    def import_file(self, file, target=None, apikey=''):
+    @property
+    def list_sites(self):
         """
-        Imports an OpenAPI definition from a local file.
+        Lists all the sites recognized by the wappalyzer addon.
         This component is optional and therefore the API will only work if it is installed
         """
-        params = {'file': file, 'apikey': apikey}
-        if target is not None:
-            params['target'] = target
-        return six.next(six.itervalues(self.zap._request(self.zap.base + 'openapi/action/importFile/', params)))
+        return six.next(six.itervalues(self.zap._request(self.zap.base + 'wappalyzer/view/listSites/')))
 
-    def import_url(self, url, hostoverride=None, apikey=''):
+    @property
+    def list_all(self):
         """
-        Imports an OpenAPI definition from a URL.
+        Lists all sites and their associated applications (technologies).
         This component is optional and therefore the API will only work if it is installed
         """
-        params = {'url': url, 'apikey': apikey}
-        if hostoverride is not None:
-            params['hostOverride'] = hostoverride
-        return six.next(six.itervalues(self.zap._request(self.zap.base + 'openapi/action/importUrl/', params)))
+        return six.next(six.itervalues(self.zap._request(self.zap.base + 'wappalyzer/view/listAll/')))
+
+    def list_site(self, site):
+        """
+        Lists all the applications (technologies) associated with a specific site.
+        This component is optional and therefore the API will only work if it is installed
+        """
+        return six.next(six.itervalues(self.zap._request(self.zap.base + 'wappalyzer/view/listSite/', {'site': site})))